fix clang warnings on osx, re #16484
