+++ Framework/CurveFitting/test/CMakeLists.txt
@@ -4,11 +4,12 @@ if ( CXXTEST_FOUND )
   include_directories ( ../../DataHandling/inc ../../TestHelpers/inc )
   # This variable is used within the cxxtest_add_test macro to build these helper classes into the test executable.
   # It will go out of scope at the end of this file so doesn't need un-setting
-  set ( TESTHELPER_SRCS ../../TestHelpers/src/TearDownWorld.cpp
+  set ( TESTHELPER_SRCS ../../TestHelpers/src/ComponentCreationHelper.cpp
                         ../../TestHelpers/src/InstrumentCreationHelper.cpp
-                        ../../TestHelpers/src/WorkspaceCreationHelper.cpp
-                        ../../TestHelpers/src/ComponentCreationHelper.cpp
+                        ../../TestHelpers/src/MultiDomainFunctionHelper.cpp
                         ../../TestHelpers/src/StartFrameworkManager.cpp
+                        ../../TestHelpers/src/TearDownWorld.cpp
+                        ../../TestHelpers/src/WorkspaceCreationHelper.cpp
   )
   cxxtest_add_test ( CurveFittingTest ${TEST_FILES}  )
   target_link_libraries( CurveFittingTest LINK_PRIVATE ${TCMALLOC_LIBRARIES_LINKTIME}
