+++ Framework/SINQ/test/PoldiSpectrumDomainFunctionTest.h
@@ -118,13 +118,11 @@ public:
     function.function(domain, values);
 
     std::array<double, 19> reference{
-        0.214381692355321, 1.4396533098854,  7.69011673999647,
-        32.6747845396612,  110.432605589092, 296.883931458002,
-        634.864220660384,  1079.89069118744, 1461.11207069126,
-        1572.50503614829,  1346.18685763306, 916.691981263516,
-        496.502218342172,  213.861997764049, 73.2741206547921,
-        19.9697293956518,  4.32910692237627, 0.746498624291666,
-        0.102391587633906};
+        {0.214381692355321, 1.4396533098854, 7.69011673999647, 32.6747845396612,
+         110.432605589092, 296.883931458002, 634.864220660384, 1079.89069118744,
+         1461.11207069126, 1572.50503614829, 1346.18685763306, 916.691981263516,
+         496.502218342172, 213.861997764049, 73.2741206547921, 19.9697293956518,
+         4.32910692237627, 0.746498624291666, 0.102391587633906}};
 
     for (size_t i = 0; i < reference.size(); ++i) {
       TS_ASSERT_DELTA(values[479 + i] / reference[i], 1.0, 1e-14);
@@ -151,13 +149,11 @@ public:
     function.functionDeriv(domain, jacobian);
 
     std::array<double, 19> reference{
-        0.214381692355321, 1.4396533098854,  7.69011673999647,
-        32.6747845396612,  110.432605589092, 296.883931458002,
-        634.864220660384,  1079.89069118744, 1461.11207069126,
-        1572.50503614829,  1346.18685763306, 916.691981263516,
-        496.502218342172,  213.861997764049, 73.2741206547921,
-        19.9697293956518,  4.32910692237627, 0.746498624291666,
-        0.102391587633906};
+        {0.214381692355321, 1.4396533098854, 7.69011673999647, 32.6747845396612,
+         110.432605589092, 296.883931458002, 634.864220660384, 1079.89069118744,
+         1461.11207069126, 1572.50503614829, 1346.18685763306, 916.691981263516,
+         496.502218342172, 213.861997764049, 73.2741206547921, 19.9697293956518,
+         4.32910692237627, 0.746498624291666, 0.102391587633906}};
 
     for (size_t i = 0; i < reference.size(); ++i) {
       TS_ASSERT_DELTA((jacobian.get(479 + i, 0)) /
